Horizontal

+++ include/UICore/Utils.h
@@ -5,10 +5,10 @@
 namespace rp::uicore
 {
     inline
-    std::string reduceNumDecimals(float value, int numDecimals)
+    std::string reduceNumDecimals(float value, size_t numDecimals)
     {
         auto stream = std::stringstream();
-        stream << std::fixed << std::setprecision(numDecimals) << value;
+        stream << std::fixed << std::setprecision(static_cast<int>(numDecimals)) << value;
         return stream.str();
     }
 }
